+++ cypress/integration/conditions/unsupportedTypes.cy.js
@@ -7,6 +7,7 @@ import {
 } from '../../data/index.js'
 import {
     openDimension,
+    openProgramDimensionsSidebar,
     selectEventWithProgram,
 } from '../../helpers/dimensions.js'
 import {
@@ -24,6 +25,8 @@ const stageName = 'Stage 1 - Repeatable'
 
 const setUpTable = () => {
     selectEventWithProgram(event)
+
+    openProgramDimensionsSidebar()
 
     selectRelativePeriod({
         label: periodLabel,
@@ -51,6 +54,8 @@ describe('unsupported types', () => {
             goToStartPage()
 
             selectEventWithProgram(E2E_PROGRAM)
+
+            openProgramDimensionsSidebar()
 
             openDimension(type.dimension)