+++ src/components/MainSidebar/MainSidebar.js
@@ -1,17 +1,28 @@
 import i18n from '@dhis2/d2-i18n'
-import { IconArrowRight16, IconFolder16 } from '@dhis2/ui'
+import { IconArrowRight16, IconFolder16, Tooltip } from '@dhis2/ui'
 import cx from 'classnames'
-import React, { useState } from 'react'
+import React from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import {
+    acSetUiAccessoryPanelActiveTab,
     acSetUiAccessoryPanelOpen,
     acSetUiDetailsPanelOpen,
 } from '../../actions/ui.js'
+import { PROGRAM_TYPE_WITH_REGISTRATION } from '../../modules/programTypes.js'
+import {
+    ACCESSORY_PANEL_TAB_INPUT,
+    ACCESSORY_PANEL_TAB_PROGRAM,
+    ACCESSORY_PANEL_TAB_YOUR,
+} from '../../modules/ui.js'
+import { OUTPUT_TYPE_EVENT } from '../../modules/visualization.js'
+import { sGetMetadataById } from '../../reducers/metadata.js'
 import {
     sGetUiInputType,
     sGetUiShowAccessoryPanel,
     sGetUiProgramId,
     sGetUiSidebarHidden,
+    sGetUiProgramStageId,
+    sGetUiAccessoryPanelActiveTab,
 } from '../../reducers/ui.js'
 import { InputPanel, getLabelForInputType } from './InputPanel/index.js'
 import { MainDimensions } from './MainDimensions.js'
@@ -22,22 +33,33 @@ import {
     SelectedDimensionsProvider,
     useSelectedDimensions,
 } from './SelectedDimensionsContext.js'
-import { TimeDimensions } from './TimeDimensions.js'
 import { YourDimensionsMenuItem } from './YourDimensionsMenuItem.js'
 import { YourDimensionsPanel } from './YourDimensionsPanel/index.js'
 
-const TAB_INPUT = 'INPUT'
-const TAB_PROGRAM = 'PROGRAM'
-const TAB_YOUR = 'YOUR'
-
 const MainSidebar = () => {
     const dispatch = useDispatch()
-    const open = useSelector(sGetUiShowAccessoryPanel)
+    const selectedTabId = useSelector(sGetUiAccessoryPanelActiveTab)
+    const open = useSelector(sGetUiShowAccessoryPanel) && Boolean(selectedTabId)
     const selectedInputType = useSelector(sGetUiInputType)
     const selectedProgramId = useSelector(sGetUiProgramId)
+    const selectedStageId = useSelector(sGetUiProgramStageId)
+    const program = useSelector((state) =>
+        sGetMetadataById(state, selectedProgramId)
+    )
+    const stage = useSelector((state) =>
+        sGetMetadataById(state, selectedStageId)
+    )
+    const subtitle =
+        selectedInputType === OUTPUT_TYPE_EVENT &&
+        program?.programType === PROGRAM_TYPE_WITH_REGISTRATION &&
+        program?.name &&
+        stage?.name
+            ? `${program.name} - ${stage.name}`
+            : program?.name
     const isHidden = useSelector(sGetUiSidebarHidden)
-    const [selectedTabId, setSelectedTabId] = useState(null)
     const setOpen = (newOpen) => dispatch(acSetUiAccessoryPanelOpen(newOpen))
+    const setSelectedTabId = (id) =>
+        dispatch(acSetUiAccessoryPanelActiveTab(id))
     const closeDetailsPanel = () => dispatch(acSetUiDetailsPanelOpen(false))
     const onClick = (id) => {
         if (open && id === selectedTabId) {
@@ -50,6 +72,17 @@ const MainSidebar = () => {
         }
     }
     const { counts } = useSelectedDimensions()
+    const programDimensionsItem = (
+        <MenuItem
+            icon={<IconFolder16 />}
+            label={i18n.t('Program dimensions')}
+            onClick={() => onClick(ACCESSORY_PANEL_TAB_PROGRAM)}
+            selected={open && selectedTabId === ACCESSORY_PANEL_TAB_PROGRAM}
+            count={counts.program}
+            disabled={!selectedProgramId}
+            dataTest="program-dimensions-button"
+        />
+    )
 
     return (
         <div className={cx(styles.container, { [styles.hidden]: isHidden })}>
@@ -60,38 +93,60 @@ const MainSidebar = () => {
                         type: getLabelForInputType(selectedInputType),
                         nsSeparator: '^^',
                     })}
-                    onClick={() => onClick(TAB_INPUT)}
-                    selected={open && selectedTabId === TAB_INPUT}
+                    onClick={() => onClick(ACCESSORY_PANEL_TAB_INPUT)}
+                    selected={
+                        open && selectedTabId === ACCESSORY_PANEL_TAB_INPUT
+                    }
+                    subtitle={subtitle}
                 />
-                <MenuItem
-                    icon={<IconFolder16 />}
-                    label={i18n.t('Program dimensions')}
-                    onClick={() => onClick(TAB_PROGRAM)}
-                    selected={open && selectedTabId === TAB_PROGRAM}
-                    count={counts.program}
-                    isCountDisabled={!selectedProgramId}
-                />
+                {!selectedProgramId ? (
+                    <Tooltip
+                        dataTest={'no-input-tooltip'}
+                        content={i18n.t('Choose an input first')}
+                        closeDelay={0}
+                        placement="bottom"
+                    >
+                        {({ onMouseOver, onMouseOut, ref }) => (
+                            <span
+                                onMouseOver={onMouseOver}
+                                onMouseOut={onMouseOut}
+                                ref={ref}
+                            >
+                                {programDimensionsItem}
+                            </span>
+                        )}
+                    </Tooltip>
+                ) : (
+                    programDimensionsItem
+                )}
+
                 <YourDimensionsMenuItem
-                    selected={open && selectedTabId === TAB_YOUR}
+                    selected={
+                        open && selectedTabId === ACCESSORY_PANEL_TAB_YOUR
+                    }
                     count={counts.your}
-                    onClick={() => onClick(TAB_YOUR)}
+                    onClick={() => onClick(ACCESSORY_PANEL_TAB_YOUR)}
                 />
                 <MainDimensions />
-                <TimeDimensions />
             </div>
             <div
                 className={cx(styles.accessory, {
                     [styles.hidden]: !open,
-                    [styles.padded]: selectedTabId === TAB_INPUT,
+                    [styles.padded]:
+                        selectedTabId === ACCESSORY_PANEL_TAB_INPUT,
                 })}
                 data-test="accessory-sidebar"
             >
                 <div className={styles.accessoryInner}>
-                    <InputPanel visible={selectedTabId === TAB_INPUT} />
-                    <ProgramDimensionsPanel
-                        visible={selectedTabId === TAB_PROGRAM}
+                    <InputPanel
+                        visible={selectedTabId === ACCESSORY_PANEL_TAB_INPUT}
                     />
-                    <YourDimensionsPanel visible={selectedTabId === TAB_YOUR} />
+                    <ProgramDimensionsPanel
+                        visible={selectedTabId === ACCESSORY_PANEL_TAB_PROGRAM}
+                    />
+                    <YourDimensionsPanel
+                        visible={selectedTabId === ACCESSORY_PANEL_TAB_YOUR}
+                    />
                 </div>
             </div>
         </div>