+++ src/modules/error.js
@@ -43,14 +43,7 @@ export const noProgramError = () =>
     visualizationError(
         EmptyBox,
         i18n.t('No program selected'),
-        i18n.t('Choose a program from the Program Dimensions sidebar.')
-    )
-
-export const noStageError = () =>
-    visualizationError(
-        EmptyBox,
-        i18n.t('No stage selected'),
-        i18n.t('Choose a stage from the Program Dimensions sidebar.')
+        i18n.t('Choose a program from the Input sidebar.')
     )
 
 export const noColumnsError = () =>